Uses Sphinx for documentation

+++ analysis/utils/astrometry.py
@@ -4,17 +4,18 @@ import pyfits
 import shutil
 from django.conf import settings
 import subprocess
+from subprocess import CalledProcessError
 
-fitscard_imagetype = 'IMAGETYP' #name of FITS Header Card that identifies the image type i.e. bias, dark, flat, science
-fitscard_filter = 'FILTER' #name of FITS Header Card that contains the filter name used in the observations
-fitscard_exptime = 'EXPTIME' #name of FITS Header Card that contains the exposure time used in the observations
-fitscard_ccdtemp = 'CCD-TEMP' #name of FITS Header Card that contains the CCD temperature used in the observations
-fitscard_objdec = 'OBJCTDEC' #name of FITS Header Card that contains the Declination of the telescope during observations
-fitscard_objra = 'OBJCTRA' #name of FITS Header Card that contains the Right Ascention of the telescope during observations
-fitscard_naxis1 = 'NAXIS1' #name of FITS Header Card that contains the number o CCD pixels along axis1
-fitscard_naxis2 = 'NAXIS2' #name of FITS Header Card that contains the number o CCD pixels along axis2
-fitscard_pierside = 'PIERSIDE' #name of FITS Header Card that contains the Pier Side of the telescope during observations
-fitscard_jd = 'JD' #name of FITS Header Card that contains the Julian Date during observations
+fitscard_imagetype = 'IMAGETYP'  # name of FITS Header Card that identifies the image type i.e. bias, dark, flat, science
+fitscard_filter = 'FILTER'  # name of FITS Header Card that contains the filter name used in the observations
+fitscard_exptime = 'EXPTIME'  # name of FITS Header Card that contains the exposure time used in the observations
+fitscard_ccdtemp = 'CCD-TEMP'  # name of FITS Header Card that contains the CCD temperature used in the observations
+fitscard_objdec = 'OBJCTDEC'  # name of FITS Header Card that contains the Declination of the telescope during observations
+fitscard_objra = 'OBJCTRA'  # name of FITS Header Card that contains the Right Ascention of the telescope during observations
+fitscard_naxis1 = 'NAXIS1'  # name of FITS Header Card that contains the number o CCD pixels along axis1
+fitscard_naxis2 = 'NAXIS2'  # name of FITS Header Card that contains the number o CCD pixels along axis2
+fitscard_pierside = 'PIERSIDE'  # name of FITS Header Card that contains the Pier Side of the telescope during observations
+fitscard_jd = 'JD'  # name of FITS Header Card that contains the Julian Date during observations
 
 
 def j(p1, p2):
@@ -75,9 +76,9 @@ def do_astrometry(path, file_id):
     # print de1, de2, de3
     # get the sign for dec and make it a multiplier
     sign = de[0][0]
-    if (sign == "+"):
+    if sign == "+":
         mult = 1.0
-    elif (sign == "-"):
+    elif sign == "-":
         mult = -1.0
     else:
         print "Something went wrong!"
@@ -90,9 +91,9 @@ def do_astrometry(path, file_id):
 
     # set pixel scale
     scale = 2.65556E-04
-    if (inhdulist[0].header[fitscard_pierside] == 'WEST'):
+    if inhdulist[0].header[fitscard_pierside] == 'WEST':
         factor = 1.0
-    elif (inhdulist[0].header[fitscard_pierside] == 'EAST'):
+    elif inhdulist[0].header[fitscard_pierside] == 'EAST':
         factor = -1.0
 
     # write the WCS keywords
@@ -103,11 +104,10 @@ def do_astrometry(path, file_id):
     inhdulist[0].header['CD2_1'] = (0.0, 'no comment')
     inhdulist[0].header['CD2_2'] = (factor * scale, 'no comment')
 
-    # print inhdulist[0].header
-
     # define the list of astrometric catalogues and bands to use
     listofcats = ['2MASS', 'USNO-B1', 'PPMX', 'NOMAD-1', 'DENIS-3', 'USNO-A1', 'USNO-A2', 'GSC-1.3', 'GSC-2.2',
                   'GSC-2.3', 'UCAC-1', 'UCAC-2', 'UCAC-3', 'SDSS-R3', 'SDSS-R5', 'SDSS-R6', 'SDSS-R7']
+
     listofbands = ['DEFAULT', 'REDDEST', 'BLUEST']
 
     # write out file and then run the astrometry software
@@ -118,8 +118,8 @@ def do_astrometry(path, file_id):
     # loop over all catalogues and bands until a working astrometric solution has been found
     cat = 0
     band = 0
-    solution_found = 0
-    while (solution_found == 0):
+    solution_found = False
+    while solution_found is False:
         catused = listofcats[cat]
         bandused = listofbands[band]
         shutil.copy(os.path.join(WORKING_DIRECTORY, "wcsprep.fits"), j(WORKING_DIRECTORY, "test.fits"))
@@ -132,7 +132,14 @@ def do_astrometry(path, file_id):
 
         scamp_command = ['scamp', j(WORKING_DIRECTORY, 'test.cat'), '-c', j(settings.CONFIGS_DIRECTORY, 'scamp.conf'),
                          '-ASTREF_CATALOG', catused, '-ASTREF_BAND', bandused]
-        scamp_output = subprocess.check_output(scamp_command)
+
+        scamp_failed = False
+        try:
+            scamp_output = subprocess.check_output(scamp_command)
+        except CalledProcessError:
+            # Wrong sky zone - no solution found...
+            scamp_failed = True
+            solution_found = False
 
         missfits_command = ['missfits', '-c', j(settings.CONFIGS_DIRECTORY, 'default.missfits'),
                             j(WORKING_DIRECTORY, 'test.fits'), '-HEADER_SUFFIX', '\'.head\'', '-WRITE_XML', 'N']
@@ -142,17 +149,18 @@ def do_astrometry(path, file_id):
         # Cleanup files dropped by sextractor and SCAMP
         os.remove(j(WORKING_DIRECTORY, 'test.fits.back'))
         os.remove(j(WORKING_DIRECTORY, 'test.cat'))
-        os.remove(j(WORKING_DIRECTORY, 'test.head'))
 
         # if whole output contains 'wrong sky zone'
         # SDSS-R3, DENIS-3 force and test
 
-        if '\x1b[01;31mtest.cat' in scamp_output:
-            print "no solution found"
-            solution_found = 0
-        else:
-            print "solution found"
-            solution_found = 1
+        if not scamp_failed:
+            os.remove(j(WORKING_DIRECTORY, 'test.head'))
+            if '\x1b[01;31mtest.cat' in scamp_output:
+                print "no solution found"
+                solution_found = False
+            else:
+                print "solution found"
+                solution_found = True
 
         if cat <= len(listofcats):
             cat = cat + 1
@@ -165,8 +173,9 @@ def do_astrometry(path, file_id):
                 catused = 'none'
                 bandused = 'none'
                 solution_found = 1
+
     # update fits header of solved image with the catalogue info
-    inhdulist = fits.get_hdu_list(j(WORKING_DIRECTORY, 'astro.fits')) # astro.fits
+    inhdulist = fits.get_hdu_list(j(WORKING_DIRECTORY, 'astro.fits'))  # astro.fits
     inhdulist[0].header['AST_CAT'] = (catused, 'Astrometric Catalogue used for WCS')
     inhdulist[0].header['AST_BAND'] = (bandused, 'Astrometric Band used for WCS')
     pyfits.writeto(j(WORKING_DIRECTORY, "astro1.fits"), inhdulist[0].data, inhdulist[0].header)
